pyproject.toml
[project]
name = "ocr-model"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "datasets>=4.5.0",
    "fastapi>=0.128.5",
    "ipykernel>=7.1.0",
    "jupyter>=1.1.1",
    "matplotlib>=3.10.8",
    "notebook>=7.5.3",
    "numpy>=2.4.2",
    "pandas>=3.0.0",
    "pillow>=12.1.0",
    "python-multipart>=0.0.22",
    "pyyaml>=6.0.3",
    "scikit-learn>=1.8.0",
    "uvicorn>=0.40.0",
]
